Handle Authentication
Kinda crap, but it works.
When we need a login, we store the request, redirect to the secret part of the app.
That part authenticates just to get to a redirect which responds to the given request.

Convoluted, but that's because of how Digest works.
I had to make a whole other class and route to it.
Beh.

Secrets are not secret as of now, which is fine.
That'll come later.